+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Implemented `pseq` by implementing `ParallelStream` and `ParallelExecutionEngine`
 * `map`, `select`, `filter`, `filter_not`, `where`, `flatten`, and `flat_map` parallelized
 * Many Thanks to `versae` for implementing the `pseq` feature!
+* Cartesian product from `itertools.product` implemented as `cartesian`
 
 ## Release 0.6.0
 ### New Features